#7 feat: MainViewModel 구현
기존 ProfileList로 관리하던 유저 정보를 viewModel을 이용해 적용
