+++ app/src/main/java/com/sopt/dive/ui/components/UserProfileInfo.kt
@@ -26,8 +26,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.sopt.dive.util.UserInfo
 import com.sopt.dive.R
+import com.sopt.dive.viewmodel.MainViewModel.UserInfo
 
 
 @Composable
