--- app/src/main/java/com/sopt/dive/util/UserProfileList.kt
@@ -1,23 +1,0 @@
-package com.sopt.dive.util
-
-data class UserInfo(val userName: String, val title: String, val content: String)
-
-val userProfileList = listOf(
-    UserInfo(userName = "지니", title = "집 보내주세요", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요2", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요3", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요4", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요5", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요6", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요7", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요8", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요9", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요10", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요11", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요12", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요13", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요14", content = "일하기 싫어요 "),
-    UserInfo(userName = "지니", title = "집 보내주세요15", content = "일하기 싫어요 "),
-
-    )
-
